Upgrade ocibuild
Newer ocibuild adds an `--output` flag to `layer gobuild` that allows
it to avoid bumping timestamps if the file contents are the same.
We've learned from the past few months of ocibuild usage that these
unnecessary timestamp bumps are a source of a lot of waiting on
`docker push`.  So switch to the `--output` flag and save some time!

Signed-off-by: Luke Shumaker <lukeshu@lukeshu.com>