Inline post-compile.sh in to Dockerfile, and simplify
OK, "inline" isn't quite the right word, since this isn't a simple
copy/paste.

post-compile.sh had 4 paragraphs:

 1. Install `busyambassador` from `/buildroot/bin` to
    `/opt/ambassador/bin`, and loop over a set of program names
    creating symlinks for each in both directories.
 2. Install `/buildroot/bin/capabilities_wrapper` to `/opt/ambassador/bin/wrapper`
 3. Run `build-aux/install.sh`.
 4. Run `post-compile.local.sh` if it exists.

For paragraphs 1 and 2, I unrolled the loop, and organized this in to
two `RUN` commands, one doing everything in `/buildroot/bin` and the
other doing everything in `/opt/ambassador/bin`.  I also shortened the
symlink targets to be relative paths instead of absolute paths.

For paragraph 3, all `install.sh` does is loop over
`build-aux/installers/` and run any scripts in that directory.  Well,
there aren't any scripts in that directory, so just drop this part.

For paragraph 4, I dropped the ability to run post-compile.local.sh.
That functionality was originally there in order for you to be able to
inject your own preferences in to the builder shell.  It was also
briefly used by apro.git to do its own extensions.  Well the builder
shell is gone and apro.git doesn't do that anymore, so just drop this
vestigial functionality.

Signed-off-by: Luke Shumaker <lukeshu@lukeshu.com>