tools/src/py-mkopensource: Implement parsing license classifiers
The "License" field (which py-mkopensource has been using) has been
deprecated for a long time; packages are supposed to use Trove
"Classifiers".  It's been working "good enough" because packages have been
slow to switch, and when they do they usually set both.  But it's becoming
more and more of a problem.

So yeah, implement parsing of license classifiers.

Signed-off-by: Luke Shumaker <lukeshu@lukeshu.com>