fix: properly handle HTTP/2 connection reuse
When multiple domains use the same certificate (e.g. the server has a certificate that can be used for domain domain.com and subdomains a.domain.com and b.domain.com) and the server supports HTTP/2, the browser will reuse the same connection for requests to domain.com, a.domain.com, and b.domain.com.

In 1.14 this wasn't an issue because all virtual hosts lived on a single filter chain.

It's a bit different in 2.y and above. To increase host and route matching performance, virtual hosts were split across multiple filter chains, using SNI for server name matching if TLS is available. For the non-TLS case, all virtual hosts are still combined under one filter chain since hostname matching in the filter_chain_match isn't available in the cleartext case.

This can an issue when reusing TLS connections. If an individual virtual_host was created for both domain.com and a.domain.com and both  were using the same TLSContext/TLS certificate, a connection would be established to the virtual_host that is first visited but that same connection will be reused when visiting the other virtual_host. Since the virtual_hosts live on separate filter chains, this results in a 404.

This updates so that Hosts sharing the same TLSContext will be coalesced into a single filter chain with both hosts added as server names for the filter_chain_match. This allows for HTTP/2 connection reuse in the browser.

If a Host and TCPMapping share the same TLSContext then the existing behavior is retained, a separate filter chain is created for each. Same for the inline and implicit tls cases.

Signed-off-by: Hamzah Qudsi <hqudsi@datawire.io>