Split integration environment setup from tests
Separates the test environment setup from the step that runs the tests.
This is to avoid conflating setup failures with test failures (it should
not require reading logs to determine whether it was the setup,
deployment, or tests that failed). This does not yet separate any test
deployment from running the tests.

Signed-off-by: Rick Lane <rlane@datawire.io>