python: consolidate pytest settings into pyproject.toml
Some of our settings were in `pytest.ini` and some were in
`pyproject.toml`.  Since we have settings for formating
and testing it makes sense to consolidate into the
pyproject.toml.

All the settings are the same with one addition which is the
`pythonpath`. This was added to make vscode happy when debugging
code.

Signed-off-by: Lance Austin <laustin@datawire.io>