deps: bump orjson and remove pyyaml from base-python
removes pyyaml from base-python because it wasn't actually being used. It is already being pip-installed with kubernetes lib and deps were set for it. It is used in kat harness so add it to requirement-devs.txt
so that it is part of the normal pip install.

Bump orjson, because newer versions have support for py 3.11 and
also have additional wheels published which make it so we do not need
to build it.

Signed-off-by: Lance Austin <laustin@datawire.io>