deps: bump code-generator to k8s v1.22
Bumps code-generator which contains `conversion-gen` the tool we use
to generate conversions between our CRD types. The OOB conversion-gen
is not able to handle all of our scenarios so cusom commits have been
place on an emissary fork.

This bumps the fork to use the latest release v1.22 for Kuberenetes so that
it matches our lowest supported version and the main version tested
for our integration/e2e tests.

Signed-off-by: Lance Austin <laustin@datawire.io>