deps: bump controller-tools to v0.7.0
This removes the custom fork of controller-tools in favor of
using upstream and matches it to the version that has k8s v1.22 libraries
as its dependencies. The additional K8s deps were updated to keep
them in-sync with each other.

We had a single commit in our fork for a generated deep-copy func.
Which addressed an issue with alias for `StringOrStringList` type used
for the `d6e-union`. This deep-copy function has been copied and
fixed and is now part of the v2.MappingSpec type and will no
longer be auto-generated. Any changes to v2.MappingSpec will
require updating this func. We should also work to see if we can get
our custom commit upstreamed.

SideNote: this StringOrStringList type has the `d6e-union`
custom tag which isn't allowed in strcutural schema's. So,
the output of controller-tools is passed along to
`tools/src/fix-crds` to address the issues and bundle all the
CRD's into a single output file.

updates kube-openapi to commit just before the switch from
googleapis/gnostic --> github.com/google/gnostic if not errors
are thrown due to type mismatches.

Signed-off-by: Lance Austin <laustin@datawire.io>