deps: bump controller-runtime to v0.10.2
This bumps controller-runtime to v0.10 which has a dep
on the k8s 1.22 libraries. This ensures controller-runtime
aligns with conversion-gen,controller-gen and the k8s dep
libraries.

The biggest change in v0.10.0 is now that it by default leverages
the v1 verion for the `conversionReviewVersions` used when K8s
talks to apiext. Adjusted `tools/src/fix-crd's` to add `v1` instead
of v1beta1 so it will now be configured to only support v1.

Signed-off-by: Lance Austin <laustin@datawire.io>