envoy: update go-control-plane to v1.26 compat
This bumps the go-control-plane version to the latest version of envoy
where the protobufs were sync'd with envoy upstream v1.26.

Bumping it, introduced a new RateLimitConfigDiscoverService which modified
the generated code/imports. These new files were not copied over and also
didn't replace the imports with the `github.com/emissary-ingress` namespaced
packages. The envoy.mk file was updated to include the new folder and
import pattern replacement

Signed-off-by: Lance Austin <laustin@datawire.io>