deps: bumps controller-tools to v0.12.0
This bumps our usage of controller-tools to v0.12.0 which is
the latest official release. Which has a dependency on the
k8s 1.27 libraries.

Some minor changes in CRD output are observed:
- x-kubernetes-map-type: atomic was added to core types.
- bumps generated by controler-gen version
- removes usage s of `\t` from descriptions
- spacing on descriptions are slightly different

`x-kubernetes-map-type: atomic` was added to a few of the underlying types
for server-side-apply. This flowed through but did require a change to tooling
as outlined here:

A bug in controller-tools and using multiple `paths...` as
outlined here: https://github.com/kubernetes-sigs/controller-tools/issues/783,
required a minor change in the input args for controller-gen. When
making this change it caused a race-condition between go:embed
in the 'pkg/api/getambassador.io/everything.go' file and the
generate-clean of the `pkg/api/getambassador.io/crds.yaml`. Therefore,
this file is no longer removed before generate runs and instead will
be overwritten by fix-crd because stdout is ">" into the file so it will overwrite
the previous contents. Once the PR is merged upstream we can revert this
back if it makes it clearer.

Signed-off-by: Lance Austin <laustin@datawire.io>