kates: log informer err when adding event handlers
In release/v1.26 of client-go the signature changed for
`informer.AddEventHandler` as seen here:
https://github.com/kubernetes/client-go/blob/release-1.26/tools/cache/shared_informer.go

This caused golangci lint to complain with:
"error return value of `informer.AddEventHandler` is not checked (errcheck)".

The new signature returns a handler reference which would allow for removing
a registered handler. Since we didn't do this and didn't in the past, we will not
need to do anything with this handle ref. However, to make golangci lint happy
we are not logging if an error occurs during registertion.

Signed-off-by: Lance Austin <laustin@datawire.io>