Replace rsync --mkpath with mkdir
Removes the use of the `rsync` `--mkpath` options in favor of creating
the target folder with `mkdir` instead. The `--mkpath` option is only
available in later versions of `rsync`, which are not generally
available on some platforms. While in most cases, it's better to rely
upon (and insist upon) the latest versions, upgrading `rsync` is
relatively prohibitive, and it would be better not to have this barrier
to open source use.

Signed-off-by: Rick Lane <rlane@datawire.io>