go-filter: replace GO_FILTER_OBJECT_FILE env var with AMBASSADOR_DISABLE_GO_FILTER
AMBASSADOR_DISABLE_GO_FILTER is an env var flag to disable injecting the go filter when running in EdgeStack mode. When running pure Emissary, the go-filter is not injected anyway. the library path is now kept hardcoded since there's not reason to really expose it to be configurable and was causing some confusion on naming.

Due to the presence of this flag now, we add it to the KAT/integration test manifests so that the go-filter isn't injected during those tests. This is fine for now because those tests are not testing any functionality from the go-filter. Future work will include overhauling the E2E test suite.

Signed-off-by: Hamzah Qudsi <hqudsi@datawire.io>