deps(go): bump go-containerregistery and indirect dep
This updates the go-containerregistery to ensure we are
using the latest `crane` tool for the build system. It also updates
transitive deps as well.

The `github.com/docker/distribution` has been updated to the latest to
quiet scanners and ensure we are on the latest.

Signed-off-by: Lance Austin <laustin@datawire.io>
(cherry picked from commit aeb71dfcedb3ce5fc51b8852d9a4504d2a555bf3)