host: add warning for unsupported previewURL field
The `Host.spec.previewURL` is a hold over from when the previewURL
feature was previously apart of the old ambassador and telepresence 1.
Documentation and support for this has long been removed.

But just in case someone is still using this and is upgrading from older
versions then we want to provide user feedback.

Signed-off-by: Lance Austin <laustin@datawire.io>