deps: bump envoy to 1.27.0-a
The existing custom build was reverted in v3.8 due to a
memory leak and seg fault that was causing Envoy to crash.

This pulls in the latest backported commits and does a clean
build. The initial load testing on this verision was successful
and I was unable to get it to seg fault.

Signed-off-by: Lance Austin <laustin@datawire.io>