envoy: update to Envoy 1.27.1 with new build process
This bumps our custom Envoy to be based on v1.27.1. The build
process has been revamped for the following:

1. align build steps with upstream Envoy's CI steps
2. Simplify envoy.mk into a set of simple Phony targets and shell scripts
3. Remove compiling protos from general `make generate`
4. Update DEVELOPING.md to match revamped workflow

A couple of key  differences are that we leverage the underlying
tools (bazel, docker) for caching and volume mounting rather than
implicit make targets. This should make it more clear what is happening
when running certain commands and will allow for more flexibility in
the dev workflow.

I tried to maintain support for FIPS_MODE but its not tested since we do not
support it and we have added it for developer. If it is not working correctly,
then follow PR's can address it as needed.

Signed-off-by: Lance Austin <laustin@dataiwre.io>