deps(python): upgrade multiple python libs via pip-compile
Ran `cd python && pip-compile --upgrade --allow-unsafe -q requirements.in`
which bumped all the deps in `python` folder.

Some notable ones:
- setuptools
- click
- charset-normalizer
- flask
- werkzeug
- typing-extensions
- gunicorn

Removed direct import of deprecated disutils because it was cause
it was causing conflicts with the latest `setuptools`.

Also, due to similiar issue as outlined here:
https://github.com/pypa/pip/issues/5247

I had to ignore uninstalling older version of packaging to allow
pip to install the newer versions in base-pip container.

Signed-off-by: Lance Austin <laustin@datawire.io>