agent: make Ambassador Agent opt-in
The Ambassador Agent is not a necessary component to
run emissary-ingress as an API Gateway. This removes it from the yaml
manifest published and turns it off by default for the Helm chart published.

We recommend you switch to the stand alone chart which can be found
here https://github.com/datawire/ambassador-agent.

Signed-off-by: Lance Austin <laustin@datawire.io>