fix-crds: add support for outputting independent files
Extends the fix-crds tool so that it can output standalone files for
RBAC, Deployments and CRD's so that we can make easier and
more predicatable e2e testing.

Generally, speaking we should re-think fix-crds and whether it makes
sense but for now this adjusts it to meet the needs for e2e testing
the new apiext server.

- removed uncessary comments that don't play nice with e2e-framework
- added port and path to default Conversion data structure to support
externally Managed mode. Note: when apiext server is patching CRD's it
will override it like it did previously.

Signed-off-by: Lance Austin <laustin@datawire.io>