apiext: add e2e test
This adds basic e2e tests for the apiext to ensure that it can properly
create, watch and renew expired Certs.

An additional test and example was added on how to run it in an externally
managed mode with CertManager providing the Certificate and Patching
CRD's. Note:  this has limited support because it has only been tested against
the same settings (RSA, PKCS8) as self-managed mode. It may work with
other settings but those are not guaranteed at this time.

A new standalone container is generated locally and pushed to k3d. This
is only for e2e testing but sets up the framework for having it standalone
in the future.

Signed-off-by: Lance Austin <laustin@datawire.io>