Revert "deps: update to envoy 1.28.0"
This reverts commit 7b7be436c59cf12f23690a5b9732416d8a5bdbff.

HTTP/3 support (udp/quic) is broken in 1.28 and will cause emissary-ingress
to shut down when Envoy tries to validate the config. In testing, 1.27 and
1.29 both are ok. So, this will temporarily revert back to Envoy 1.27.2
until 1.29 is released and we can jump to 1.29 instead.

Note: none of the current commits on the unreleased 1.28.1
branch (release/v1.28) seem to address this but rather quite a few
larger commits in master may be required thus not macking it feasible
to backport.

Signed-off-by: Lance Austin <laustin@datawire.io>